Don't auto clean before VSCode build
